pyproject.toml
[project]
name = "awesome-failed-ml-experiments"
version = "1.0.0"
description = "A curated collection of real machine learning experiments that didn't work out"
authors = [
    {name = "Ritesh Rana", email = "riteshrana36@gmail.com"}
]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
keywords = ["machine-learning", "failures", "awesome-list", "learning"]

dependencies = [
    "requests>=2.31.0",
    "pyyaml>=6.0",
    "python-frontmatter>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "black>=24.0.0",
    "ruff>=0.1.0",
    "pytest>=8.0.0",
]

[project.urls]
Homepage = "https://github.com/ambicuity/awesome-failed-ml-experiments"
Repository = "https://github.com/ambicuity/awesome-failed-ml-experiments"
Issues = "https://github.com/ambicuity/awesome-failed-ml-experiments/issues"

[tool.black]
line-length = 100
target-version = ["py311"]
include = '\.pyi?$'
exclude = '''
/(
    \.git
    | \.venv
    | __pycache__
    | build
    | dist
)/
'''

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by black)
    "B008",   # do not perform function calls in argument defaults
]

[tool.ruff.lint.isort]
known-first-party = ["scripts"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
